Move doc_events out of API folder
