+++ next_crm/api/todo.py
@@ -4,42 +4,6 @@ from frappe import _
 from next_crm.ncrm.doctype.crm_notification.crm_notification import notify_user
 
 
-def after_insert(doc, method):
-    if (
-        doc.reference_type in ["Lead", "Opportunity"]
-        and doc.reference_name
-        and doc.allocated_to
-    ):
-        fieldname = (
-            "lead_owner" if doc.reference_type == "Lead" else "opportunity_owner"
-        )
-        lead_owner = frappe.db.get_value(
-            doc.reference_type, doc.reference_name, fieldname
-        )
-        if not lead_owner:
-            frappe.db.set_value(
-                doc.reference_type, doc.reference_name, fieldname, doc.allocated_to
-            )
-
-    if (
-        doc.reference_type in ["Lead", "Opportunity", "ToDo"]
-        and doc.reference_name
-        and doc.allocated_to
-    ):
-        notify_assigned_user(doc)
-
-
-def on_update(doc, method):
-    if (
-        doc.has_value_changed("status")
-        and doc.status == "Cancelled"
-        and doc.reference_type in ["Lead", "Opportunity", "ToDo"]
-        and doc.reference_name
-        and doc.allocated_to
-    ):
-        notify_assigned_user(doc, is_cancelled=True)
-
-
 def notify_assigned_user(doc, is_cancelled=False):
     _doc = frappe.get_doc(doc.reference_type, doc.reference_name)
     owner = frappe.get_cached_value("User", frappe.session.user, "full_name")
@@ -133,17 +97,3 @@ def get_redirect_to_doc(doc):
         return reference_doc.reference_type, reference_doc.reference_name
 
     return doc.reference_type, doc.reference_name
-
-
-def before_insert(doc, method=None):
-    from frappe import get_value
-    from frappe.desk.doctype.notification_log.notification_log import get_title
-
-    if not doc.custom_title and (doc.reference_type and doc.reference_name):
-        title = get_title(doc.reference_type, doc.reference_name)
-        doc.custom_title = title
-
-    if not doc.reference_type == "Task":
-        return
-    ref_doc_desc = get_value(doc.reference_type, doc.reference_name, "description")
-    doc.description = ref_doc_desc
