+++ next_crm/doc_events/comment.py
@@ -1,0 +1,5 @@
+from next_crm.api.comment import notify_mentions
+
+
+def on_update(doc, method=None):
+    notify_mentions(doc)
